Simplify handlers.HealthHandler
The current implementation declares a Health variable in main.go like this:

    var Health struct {
            Lock   sync.Mutex
            Status int
    }

After initialization Health.Status is set to 0 (int's zero value)
and Health.Lock Mutex is initialized.

Health is then passed as argument to:

	mux.Handle("/health", handlers.HealthHandler(&Health, &Health.Status, &Health.Lock))

However Health is never used anywhere else thus Health.Status is always 0 as
well as never updated. As such the mutex around Health.Status is useless.

Furthermore in healt.HealthHandler we have this:

    [...]

    if *status == 0 {
		*status = http.StatusOK
	}

    [...]

But since Health.Status is always 0 the response code returned in the JSON
response will always be "200 OK".

The current implementation returns the following JSON:

    ➜  simple-builder git:(cleanup) ✗ http localhost:8000/health
    HTTP/1.1 200 OK
    Content-Length: 25
    Content-Type: application/json
    Date: Tue, 06 Aug 2019 10:37:14 GMT

    {
        "Lock": {},
        "Status": 200
    }

Lock should never be exported to JSON because :

    1. this is useless internal stuff as demonstrated above

    2. sync.Mutex has not exportable fields (https://golang.org/src/sync/mutex.go?s=765:813#L15)

All in all the /health endpoint should simply return "200 OK" as it actually
always does but in a super complicated way.

That's why I simplified this handler quite aggressively.